+++ test/ios/Classes/AdjustCommandExecutor.h
@@ -3,7 +3,7 @@
 //  Adjust SDK
 //
 //  Created by Srdjan Tubin (@2beens) on 1st October 2018.
-//  Copyright (c) 2018 Adjust GmbH. All rights reserved.
+//  Copyright (c) 2018-2021 Adjust GmbH. All rights reserved.
 //
 
 #import <Flutter/Flutter.h>
@@ -11,8 +11,10 @@
 #import <AdjustTestLibrary/ATLTestLibrary.h>
 
 @interface AdjustCommandExecutor : NSObject<AdjustCommandDelegate>
+
 - (id)initWithMethodChannel:(FlutterMethodChannel *)channel;
 - (void)executeCommand:(NSString *)className
             methodName:(NSString *)methodName
             parameters:(NSString *)jsonParameters;
+
 @end